% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmscorev.R
\name{groupmscorev}
\alias{groupmscorev}
\title{Calculate Group M-Score}
\usage{
groupmscorev(raw, cmat, inner_para = 0, trim_para = 3, lambda = 0.5)
}
\arguments{
\item{raw}{A numeric vector representing the raw scores.}

\item{cmat}{A matrix representing the group categories.}

\item{inner_para}{A numeric value for the inner parameter. Default is 0.}

\item{trim_para}{A numeric value for the trim parameter. Default is 3.}

\item{lambda}{A numeric value for the lambda parameter. Default is 0.5.}
}
\value{
A numeric vector of trimmed scores.
}
\description{
groupmscorev makes subgroup trimming for submax method.
}
\examples{
# 1. Generate baseline parameters
set.seed(123) # Set seed for reproducibility

#Baseline parameters
I = 1000
c = rep(1,I)
c1 = c(rep(1,I/2),rep(0,I*1/2))
c1_rev = 1-c1
c2 = rep(c(0,1),I/2)
c2_rev = 1 - c2

# Create the cmat matrix
cmat = cbind(c,c1,c1_rev,c2,c2_rev)
# Simulate the raw scores
effect_1 = 4
effect_2 = 0.2
raw = rep(0,I)
raw[1:(I/2)] = 5*rnorm(n = I/2)+effect_1
raw[(1+I/2):I] = rnorm(n = I/2)+effect_2
# 2. Generate baseline parameters
trimmed_default <- groupmscorev(raw,cmat)

}
